reproject(): Handle dst_bounds correctly when dst_res provided.
Previously, this was erroneously being handed off to crop().

Now, if dst_bounds and dst_res provided, then the new raster size is calculated in-method, ensuring that pixel size remains fixed to dst_res. This means that dst_bounds will be adjusted (increased slightly) if it was not already divisible into the requested pixel size.