Dealing with VRT read/write problems.
Read: rasterio seemingly cannot cope with opening a VRT file through its MemoryFile functionality. I am no longer sure of the benefits of opening all files through a MemoryFile as they are write-protected via open(...,'r') and all subseuqent modification operations occur upon a newly created MemoryFile. So, for now, __init__ now has a new kwarg as_memfile=False. This solves VRT opening and reading problems.

Write: I encountered this problem when trying to crop a VRT file. In this process, the driver of the existing dataset - VRT - is passed in the metadata through to _update(), which then fails as rasterio does not support writing data to VRTs. So, _update() now automatically detects attempted VRT update and in these cases defaults to the GTiff driver for the new MemoryFile which always results from this _update() operation.